requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
trafilatura>=1.9.0
pyyaml>=6.0
jinja2>=3.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
respx>=0.21.0
soundfile>=0.12.0
